refactor: split DosBatchEngine into partials
